refactor(web): split large web frontend surfaces
