+++ apps/web/src/hooks/chat/derived/use-web-cloud-transcript-projection.ts
@@ -32,8 +32,10 @@ import { resolveCloudTranscriptSessionViewState } from "../../../lib/domain/chat
 import type { PendingHomePrompt } from "../../../lib/access/cloud/pending-home-prompt-store";
 import type {
   WebCloudPromptIntent,
+} from "../../../stores/cloud/web-cloud-prompt-intent-store";
+import type {
   WebCloudSessionDraft,
-} from "../../../stores/cloud/web-cloud-chat-state-store";
+} from "../../../stores/cloud/web-cloud-session-draft-store";
 
 export function useWebCloudTranscriptProjection(input: {
   workspace: CloudWorkspaceDetail | null;
