+++ apps/web/src/hooks/chat/facade/build-web-cloud-chat-surface-props.ts
@@ -51,7 +51,7 @@ import {
   webCloudSessionDraftOptionId,
   webCloudSessionDraftSearch,
   type WebCloudSessionDraft,
-} from "../../../stores/cloud/web-cloud-chat-state-store";
+} from "../../../stores/cloud/web-cloud-session-draft-store";
 
 export function buildWebCloudChatSurfaceProps(input: {
   workspace: CloudWorkspaceDetail;
