+++ apps/web/src/hooks/chat/lifecycle/use-web-cloud-chat-local-state-lifecycle.ts
@@ -17,16 +17,20 @@ import {
   type PendingHomePrompt,
 } from "../../../lib/access/cloud/pending-home-prompt-store";
 import {
-  clearWebCloudSessionDraft,
-  loadWebCloudPendingConfigChanges,
   loadWebCloudPromptIntents,
-  loadWebCloudSessionDraftFromSearch,
-  saveWebCloudPendingConfigChanges,
   saveWebCloudPromptIntents,
-  saveWebCloudSessionDraft,
   type WebCloudPromptIntent,
+} from "../../../stores/cloud/web-cloud-prompt-intent-store";
+import {
+  loadWebCloudPendingConfigChanges,
+  saveWebCloudPendingConfigChanges,
+} from "../../../stores/cloud/web-cloud-pending-config-change-store";
+import {
+  clearWebCloudSessionDraft,
+  loadWebCloudSessionDraftFromSearch,
+  saveWebCloudSessionDraft,
   type WebCloudSessionDraft,
-} from "../../../stores/cloud/web-cloud-chat-state-store";
+} from "../../../stores/cloud/web-cloud-session-draft-store";
 
 export type PendingHomePromptDispatchRun = {
   key: string;
