+++ apps/web/src/hooks/chat/lifecycle/use-web-cloud-pending-home-prompt-lifecycle.ts
@@ -21,10 +21,12 @@ import {
   type PendingHomePrompt,
 } from "../../../lib/access/cloud/pending-home-prompt-store";
 import {
-  clearWebCloudSessionDraft,
   type WebCloudPromptIntent,
+} from "../../../stores/cloud/web-cloud-prompt-intent-store";
+import {
+  clearWebCloudSessionDraft,
   type WebCloudSessionDraft,
-} from "../../../stores/cloud/web-cloud-chat-state-store";
+} from "../../../stores/cloud/web-cloud-session-draft-store";
 import type { PendingHomePromptDispatchRun } from "./use-web-cloud-chat-local-state-lifecycle";
 
 type DispatchPendingHomePromptArgs = Parameters<typeof dispatchPendingHomePrompt>[0];
