+++ apps/web/src/hooks/chat/workflows/use-web-cloud-chat-actions.ts
@@ -23,8 +23,10 @@ import type {
 import type { PendingHomePrompt } from "../../../lib/access/cloud/pending-home-prompt-store";
 import type {
   WebCloudPromptIntent,
+} from "../../../stores/cloud/web-cloud-prompt-intent-store";
+import type {
   WebCloudSessionDraft,
-} from "../../../stores/cloud/web-cloud-chat-state-store";
+} from "../../../stores/cloud/web-cloud-session-draft-store";
 import { useWebCloudPlanDecisionActions } from "./use-web-cloud-plan-decision-actions";
 import { useWebCloudPromptActions } from "./use-web-cloud-prompt-actions";
 import { useWebCloudSessionConfigActions } from "./use-web-cloud-session-config-actions";
