+++ apps/web/src/stores/cloud/web-cloud-state-storage.ts
@@ -1,0 +1,42 @@
+export function readSessionStorageValue(key: string): string | null {
+  try {
+    return window.sessionStorage.getItem(key);
+  } catch {
+    return null;
+  }
+}
+
+export function writeSessionStorageValue(key: string, value: string): void {
+  try {
+    window.sessionStorage.setItem(key, value);
+  } catch {
+    // The in-memory copy still carries state across same-tab navigation.
+  }
+}
+
+export function removeSessionStorageValue(key: string): void {
+  try {
+    window.sessionStorage.removeItem(key);
+  } catch {
+    // Nothing to clear if storage is unavailable.
+  }
+}
+
+export function stringOrNull(value: unknown): string | null {
+  return typeof value === "string" && value.trim() ? value.trim() : null;
+}
+
+export function numberOrDefault(value: unknown, fallback: number): number {
+  return typeof value === "number" && Number.isFinite(value) ? value : fallback;
+}
+
+export function isFresh(createdAt: number, maxAgeMs: number): boolean {
+  return Date.now() - createdAt < maxAgeMs;
+}
+
+export function randomSuffix(): string {
+  if (typeof crypto !== "undefined" && "randomUUID" in crypto) {
+    return crypto.randomUUID().slice(0, 8);
+  }
+  return Math.random().toString(36).slice(2, 10);
+}
